requirements.txt
PyYAML==3.11
requests==2.8.1
python-logstash==0.4.5
